+++ config.py
@@ -1,0 +1,5 @@
+## 設定ファイルもどき
+TOKEN = 'BOT_TOKEN_DETH'
+GUILD_ID = 1255359848644608035
+CHANNEL_ID = 1255362057709289493
+TARGET_MEMBER_COUNT = 1000
